requirements.txt
# This is a requirements file for a Python project.
streamlit
Pillow
numpy

